+++ src/lib.rs
@@ -63,10 +63,12 @@ mod sysreg;
 
 pub use safe_mmio::UniqueMmioPointer;
 use safe_mmio::fields::ReadPureWrite;
+use thiserror::Error;
+use zerocopy::{FromZeros, Immutable, IntoBytes, KnownLayout};
 
 #[cfg(all(target_arch = "aarch64", not(feature = "fakes")))]
 use core::arch::asm;
-use core::fmt::{Debug, Formatter, Result};
+use core::fmt::{self, Debug, Formatter};
 
 /// The trigger configuration for an interrupt.
 #[derive(Copy, Clone, Debug, Eq, PartialEq)]
@@ -89,7 +91,10 @@ pub enum InterruptGroup {
 }
 
 /// An interrupt ID.
-#[derive(Copy, Clone, Eq, Ord, PartialOrd, PartialEq)]
+#[derive(
+    Copy, Clone, Eq, FromZeros, Immutable, IntoBytes, KnownLayout, Ord, PartialOrd, PartialEq,
+)]
+#[repr(transparent)]
 pub struct IntId(u32);
 
 impl IntId {
@@ -123,6 +128,9 @@ impl IntId {
     /// The maximum number of extended Shared Peripheral Interrupts which may be supported.
     pub const MAX_ESPI_COUNT: u32 = Self::ESPI_END - Self::ESPI_START;
 
+    /// The maximum number of Locality-specific Peripheral Interrupts which may be supported.
+    pub const MAX_LPI_COUNT: u32 = Self::LPI_END - Self::LPI_START;
+
     /// The ID of the first Software Generated Interrupt.
     const SGI_START: u32 = 0;
 
@@ -152,6 +160,9 @@ impl IntId {
 
     /// The first Locality-specific Peripheral Interrupt.
     const LPI_START: u32 = 8192;
+
+    /// One more than the last possible Locality-specific Peripheral Interrupt.
+    const LPI_END: u32 = 1 << 24;
 
     /// Returns the interrupt ID for the given Software Generated Interrupt.
     pub const fn sgi(sgi: u32) -> Self {
@@ -185,6 +196,7 @@ impl IntId {
 
     /// Returns the interrupt ID for the given Locality-specific Peripheral Interrupt.
     pub const fn lpi(lpi: u32) -> Self {
+        assert!(lpi < Self::MAX_LPI_COUNT);
         Self(Self::LPI_START + lpi)
     }
 
@@ -258,6 +270,13 @@ impl IntId {
         }
     }
 
+    /// Returns the raw interrupt ID value.
+    ///
+    /// This is the same as using the `From<IntId>` implementation, but const.
+    pub const fn raw_value(self) -> u32 {
+        self.0
+    }
+
     // TODO: Change this to return a Range<IntId> once core::iter::Step is stabilised.
     /// Returns an array of all interrupt Ids that are private to a core, i.e. SGIs and PPIs.
     pub fn private() -> impl Iterator<Item = IntId> {
@@ -275,7 +294,7 @@ impl IntId {
 }
 
 impl Debug for IntId {
-    fn fmt(&self, f: &mut Formatter) -> Result {
+    fn fmt(&self, f: &mut Formatter) -> fmt::Result {
         if self.0 < Self::PPI_START {
             write!(f, "SGI {}", self.0 - Self::SGI_START)
         } else if self.0 < Self::SPI_START {
@@ -303,6 +322,26 @@ impl Debug for IntId {
 impl From<IntId> for u32 {
     fn from(intid: IntId) -> Self {
         intid.0
+    }
+}
+
+#[derive(Clone, Copy, Debug, Eq, PartialEq, Error)]
+#[error("Invalid (reserved) interrupt ID {0}")]
+pub struct InvalidIntId(u32);
+
+impl TryFrom<u32> for IntId {
+    type Error = InvalidIntId;
+
+    fn try_from(value: u32) -> Result<Self, Self::Error> {
+        if (Self::SPECIAL_END..Self::EPPI_START).contains(&value)
+            || (Self::EPPI_END..Self::ESPI_START).contains(&value)
+            || (Self::ESPI_END..Self::LPI_START).contains(&value)
+            || value >= Self::LPI_END
+        {
+            Err(InvalidIntId(value))
+        } else {
+            Ok(Self(value))
+        }
     }
 }